+++ plugins/tflite_flutter_plugin/lib/src/interpreter.dart
@@ -20,8 +20,11 @@ class Interpreter {
   bool _allocated = false;
   int _lastNativeInferenceDurationMicroSeconds = 0;
 
-  List<Tensor> _inputTensors;
-  List<Tensor> _outputTensors;
+  List<Tensor>? _inputTensors;
+  List<Tensor>? _outputTensors;
+
+  int? _inputTensorsCount;
+  int? _outputTensorsCount;
 
   int get lastNativeInferenceDurationMicroSeconds =>
       _lastNativeInferenceDurationMicroSeconds;
@@ -34,7 +37,7 @@ class Interpreter {
   /// Creates interpreter from model
   ///
   /// Throws [ArgumentError] is unsuccessful.
-  factory Interpreter._create(Model model, {InterpreterOptions options}) {
+  factory Interpreter._create(Model model, {InterpreterOptions? options}) {
     final interpreter = tfLiteInterpreterCreate(
         model.base, options?.base ?? cast<TfLiteInterpreterOptions>(nullptr));
     checkArgument(isNotNull(interpreter),
@@ -62,7 +65,7 @@ class Interpreter {
   ///   return fileOnDevice;
   /// }
   /// ```
-  factory Interpreter.fromFile(File modelFile, {InterpreterOptions options}) {
+  factory Interpreter.fromFile(File modelFile, {InterpreterOptions? options}) {
     final model = Model.fromFile(modelFile.path);
     final interpreter = Interpreter._create(model, options: options);
     model.delete();
@@ -86,7 +89,7 @@ class Interpreter {
   ///   }
   /// ```
   factory Interpreter.fromBuffer(Uint8List buffer,
-      {InterpreterOptions options}) {
+      {InterpreterOptions? options}) {
     final model = Model.fromBuffer(buffer);
     final interpreter = Interpreter._create(model, options: options);
     model.delete();
@@ -95,7 +98,8 @@ class Interpreter {
 
   /// Creates interpreter from a [assetName]
   ///
-  /// Place your `.tflite` file in the assets folder.
+  /// Place your `.tflite` file in the assets folder. Do not include the "assets/"
+  /// directory in assetName.
   ///
   /// Example:
   ///
@@ -103,26 +107,14 @@ class Interpreter {
   /// final interpreter = await tfl.Interpreter.fromAsset('your_model.tflite');
   /// ```
   static Future<Interpreter> fromAsset(String assetName,
-      {InterpreterOptions options}) async {
-    Uint8List buffer;
-    try {
-      buffer = await _getBuffer(assetName);
-    } catch (err) {
-      print(
-          'Caught error: $err, while trying to create interpreter from asset: $assetName');
-    }
+      {InterpreterOptions? options}) async {
+    Uint8List buffer = await _getBuffer(assetName);
     return Interpreter.fromBuffer(buffer, options: options);
   }
 
   /// Get byte buffer
   static Future<Uint8List> _getBuffer(String assetFileName) async {
-    ByteData rawAssetFile;
-    try {
-      rawAssetFile = await rootBundle.load('assets/$assetFileName');
-    } catch (err) {
-      print(
-          'Caught error: $err, while trying to load asset from "assets/$assetFileName"');
-    }
+    ByteData rawAssetFile = await rootBundle.load('assets/$assetFileName');
     final rawBytes = rawAssetFile.buffer.asUint8List();
     return rawBytes;
   }
@@ -167,10 +159,10 @@ class Interpreter {
 
   /// Run for multiple inputs and outputs
   void runForMultipleInputs(List<Object> inputs, Map<int, Object> outputs) {
-    if (inputs == null || inputs.isEmpty) {
+    if (inputs.isEmpty) {
       throw ArgumentError('Input error: Inputs should not be null or empty.');
     }
-    if (outputs == null || outputs.isEmpty) {
+    if (outputs.isEmpty) {
       throw ArgumentError('Input error: Outputs should not be null or empty.');
     }
 
@@ -201,14 +193,14 @@ class Interpreter {
 
     var outputTensors = getOutputTensors();
     for (var i = 0; i < outputTensors.length; i++) {
-      outputTensors[i].copyTo(outputs[i]);
+      outputTensors[i].copyTo(outputs[i]!);
     }
   }
 
   /// Gets all input tensors associated with the model.
   List<Tensor> getInputTensors() {
     if (_inputTensors != null) {
-      return _inputTensors;
+      return _inputTensors!;
     }
 
     var tensors = List.generate(
@@ -222,7 +214,7 @@ class Interpreter {
   /// Gets all output tensors associated with the model.
   List<Tensor> getOutputTensors() {
     if (_outputTensors != null) {
-      return _outputTensors;
+      return _outputTensors!;
     }
 
     var tensors = List.generate(
@@ -236,12 +228,12 @@ class Interpreter {
   /// Resize input tensor for the given tensor index. `allocateTensors` must be called again afterward.
   void resizeInputTensor(int tensorIndex, List<int> shape) {
     final dimensionSize = shape.length;
-    final dimensions = allocate<Int32>(count: dimensionSize);
+    final dimensions = calloc<Int32>(dimensionSize);
     final externalTypedData = dimensions.asTypedList(dimensionSize);
     externalTypedData.setRange(0, dimensionSize, shape);
     final status = tfLiteInterpreterResizeInputTensor(
         _interpreter, tensorIndex, dimensions, dimensionSize);
-    free(dimensions);
+    calloc.free(dimensions);
     checkState(status == TfLiteStatus.ok);
     _inputTensors = null;
     _outputTensors = null;
@@ -250,11 +242,14 @@ class Interpreter {
 
   /// Gets the input Tensor for the provided input index.
   Tensor getInputTensor(int index) {
+    if (_inputTensorsCount == null) {
+      _inputTensorsCount = tfLiteInterpreterGetInputTensorCount(_interpreter);
+    }
+    if (index < 0 || index >= _inputTensorsCount!) {
+      throw ArgumentError('Invalid input Tensor index: $index');
+    }
     if (_inputTensors != null) {
-      if (index < 0 || index > _inputTensors.length) {
-        throw ArgumentError('Invalid input Tensor index: $index');
-      }
-      return _inputTensors[index];
+      return _inputTensors![index];
     }
 
     final inputTensor =
@@ -264,13 +259,15 @@ class Interpreter {
 
   /// Gets the output Tensor for the provided output index.
   Tensor getOutputTensor(int index) {
-    if (_outputTensors != null) {
-      if (index < 0 || index > _outputTensors.length) {
-        throw ArgumentError('Invalid output Tensor index: $index');
-      }
-      return _outputTensors[index];
+    if (_outputTensorsCount == null) {
+      _outputTensorsCount = tfLiteInterpreterGetOutputTensorCount(_interpreter);
     }
-
+    if (index < 0 || index >= _outputTensorsCount!) {
+      throw ArgumentError('Invalid output Tensor index: $index');
+    }
+    if (_outputTensors != null) {
+      return _outputTensors![index];
+    }
     final outputTensor =
         Tensor(tfLiteInterpreterGetOutputTensor(_interpreter, index));
     return outputTensor;
@@ -284,7 +281,7 @@ class Interpreter {
       inputTensorsIndex[inputTensors[i].name] = i;
     }
     if (inputTensorsIndex.containsKey(opName)) {
-      return inputTensorsIndex[opName];
+      return inputTensorsIndex[opName]!;
     } else {
       throw ArgumentError(
           "Input error: $opName' is not a valid name for any input. Names of inputs and their indexes are $inputTensorsIndex");
@@ -299,7 +296,7 @@ class Interpreter {
       outputTensorsIndex[outputTensors[i].name] = i;
     }
     if (outputTensorsIndex.containsKey(opName)) {
-      return outputTensorsIndex[opName];
+      return outputTensorsIndex[opName]!;
     } else {
       throw ArgumentError(
           "Output error: $opName' is not a valid name for any output. Names of outputs and their indexes are $outputTensorsIndex");