+++ plugins/tflite_flutter_plugin/lib/tflite_flutter.dart
@@ -13,12 +13,13 @@ export 'src/bindings/types.dart';
 export 'src/delegate.dart';
 export 'src/delegates/gpu_delegate.dart';
 export 'src/delegates/metal_delegate.dart';
-export 'src/delegates/nnapi_delegate.dart';
+export 'src/delegates/xnnpack_delegate.dart';
 export 'src/interpreter.dart';
 export 'src/interpreter_options.dart';
 export 'src/quanitzation_params.dart';
 export 'src/tensor.dart';
 export 'src/util/list_shape_extension.dart';
+export 'src/util/byte_conversion_utils.dart';
 
 /// tflite version information.
-String get version => Utf8.fromUtf8(tfLiteVersion());
+String get version => tfLiteVersion().toDartString();